+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Updates
 * expand IPSet argument type from `Sequence[str]` to `Iterable[str]` for better compatibility with type checking tools and synchronization with actual accepted types
+* Fix: error if `__neq__` is called with non IPSet object
 
 
 ## v0.1.1 - 2025-05-08